refactor(fuel-streams): removing streams for sdk
