+++ crates/fuel-streams/src/lib.rs
@@ -1,11 +1,9 @@
 pub mod client;
 mod error;
 pub mod stream;
-mod streams;
 
 pub use error::*;
 pub use stream::*;
-pub use streams::*;
 
 pub mod prelude {
     pub use crate::{client::*, error::*, stream::*};
