+++ tests/src/lib.rs
@@ -1,7 +1,28 @@
 use std::time::Duration;
 
-use fuel_streams::Streams;
-use fuel_streams_core::prelude::*;
+use fuel_streams_core::{
+    nats::NatsClient,
+    prelude::*,
+    types::{Block, Transaction},
+    Stream,
+};
+
+#[derive(Debug, Clone)]
+pub struct Streams {
+    pub blocks: Stream<Block>,
+    pub transactions: Stream<Transaction>,
+}
+
+impl Streams {
+    pub async fn new(client: &NatsClient) -> Self {
+        let blocks = Stream::<Block>::get_or_init(client).await;
+        let transactions = Stream::<Transaction>::get_or_init(client).await;
+        Self {
+            transactions,
+            blocks,
+        }
+    }
+}
 
 pub async fn server_setup() -> BoxedResult<(NatsClient, Streams)> {
     let opts = NatsClientOpts::admin_opts(NATS_URL).with_rdn_namespace();
